requirements.txt
alembic==1.0.8
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
googlemaps==4.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
requests==2.27.1
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.26.9
Werkzeug==0.16.1
WTForms==2.2.1
boto3==1.24.3
botocore==1.27.3